central-api/requirements.txt
# === Core Framework ===
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0

# === Database ===
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# === Cache & Queue ===
redis==5.0.1
aioredis==2.0.1

# === Security & Crypto ===
cryptography==41.0.8
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
PyJWT==2.8.0

# === HTTP Client ===
httpx==0.25.2
aiohttp==3.9.1

# === Monitoring & Metrics ===
prometheus-client==0.19.0
structlog==23.2.0

# === Rate Limiting ===
slowapi==0.1.9

# === Validation ===
validators==0.22.0

# === Development ===
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# === Production ===
gunicorn==21.2.0